% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{fitBBox}
\alias{fitBBox}
\title{Set Bounds of a Map Widget}
\usage{
fitBBox(map, bbox)
}
\arguments{
\item{map}{a map widget object created from \code{\link[leaflet]{leaflet}()}}

\item{bbox}{A vector \code{c(lng1, lat1, lng2, lat2)} specifying the bounding box
coordinates}
}
\value{
The modified map widget.
}
\description{
Helper function to set the bounds of a leaflet map widget.
}
\author{
Andrzej Oleś \href{mailto:andrzej.oles@gmail.com}{andrzej.oles@gmail.com}
}
