% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{ors_directions}
\alias{ors_directions}
\title{Openrouteservice Directions}
\usage{
ors_directions(coordinates, profile = c("driving-car", "driving-hgv",
  "cycling-regular", "cycling-road", "cycling-safe", "cycling-mountain",
  "cycling-tour", "cycling-electric", "foot-walking", "foot-hiking",
  "wheelchair"), ...)
}
\arguments{
\item{coordinates}{List of \code{longitude, latitude} coordinate pairs visited in
order.}

\item{profile}{Route profile, defaults to \code{driving-car}.}

\item{...}{Optional parameters as described \href{https://openrouteservice.org/documentation/#/reference/directions/directions}{here}.}
}
\value{
Route between two or more locations for a selected profile and its
settings as GeoJSON response.
}
\description{
Get directions for different modes of transport.
}
\examples{
coordinates = list(c(8.34234, 48.23424), c(8.34423, 48.26424))

# simple call
ors_directions(coordinates, preference="fastest")

#customized options
ors_directions(coordinates, profile="cycling-mountain", elevation=TRUE, format="geojson")
}
\author{
Andrzej Oleś <\email{andrzej@openrouteservice.org}>
}
