% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones.R
\name{ors_isochrones}
\alias{ors_isochrones}
\title{Openrouteservice Isochrones}
\usage{
ors_isochrones(locations, profile = ors_profile(), range = 60, ...,
  api_key = ors_api_key(), parse_output = NULL)
}
\arguments{
\item{locations}{List of \code{longitude, latitude} coordinate pairs, alternatively a two column \code{matrix} or \code{data.frame}.}

\item{profile}{Route profile, defaults to \code{"driving-car"}.}

\item{range}{Maximum range value of the analysis in seconds for time and
meters for distance. Alternatively a comma separated list of specific
single range values.}

\item{...}{Optional parameters as described \href{https://openrouteservice.org/dev/#/api-docs/isochrones}{here}.}

\item{api_key}{Character scalar containing openrouteservice API key}

\item{parse_output}{Should the response be parsed to an R object}
}
\value{
A GeoJSON object containing a FeatureCollection of Polygons
corresponding to the accessible area.
}
\description{
Obtain areas of reachability from given locations.
}
\details{
The Isochrone Service supports time and distance analyses for one single or
multiple locations. You may also specify the isochrone interval or provide
multiple exact isochrone range values.
}
\examples{
ors_isochrones(c(8.34234, 48.23424), interval=20)

locations <- list(c(8.681495, 49.41461), c(8.686507,49.41943))
ors_isochrones(locations, range=c(300, 200))
}
\author{
Andrzej Oleś <\email{andrzej@openrouteservice.org}>
}
