% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{ors_matrix}
\alias{ors_matrix}
\title{Openrouteservice Matrix}
\usage{
ors_matrix(locations, profile = c("driving-car", "driving-hgv",
  "cycling-regular", "cycling-road", "cycling-safe", "cycling-mountain",
  "cycling-tour", "cycling-electric", "foot-walking", "foot-hiking",
  "wheelchair"), metrics = c("distance", "duration"), ...,
  parse_output = NULL)
}
\arguments{
\item{locations}{List of \code{longitude, latitude} coordinate pairs}

\item{profile}{Route profile, defaults to \code{"driving-car"}.}

\item{metrics}{Returned metrics. Use \code{"distance"} for distance matrix in defined \code{units}, and/or `duration`` for time matrix in seconds.}

\item{...}{Optional parameters as described \href{https://openrouteservice.org/documentation/#/reference/matrix/matrix}{here}.}

\item{parse_output}{Should the response be parsed to an R object}
}
\value{
Duration or distance matrix for mutliple source and destination points.
}
\description{
Obtain one-to-many, many-to-one and many-to-many matrices for time and distance.
}
\examples{
coordinates = list(
  c(9.970093, 48.477473),
  c(9.207916, 49.153868),
  c(37.573242, 55.801281),
  c(115.663757,38.106467)
)

# query for duration and distance in km
res = ors_matrix(coordinates, metrics = c("duration", "distance"), units = "km")

# duration in hours
res$durations / 3600

# distance in km
res$distances
}
\author{
Andrzej Oleś <\email{andrzej@openrouteservice.org}>
}
