% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois.R
\name{ors_pois}
\alias{ors_pois}
\title{Openrouteservice POIs}
\usage{
ors_pois(request = c("pois", "stats", "list"), geometry, ...,
  api_key = ors_api_key(), output = c("parsed", "text"))
}
\arguments{
\item{request}{One of the following: \code{"pois"}, \code{"stats"} or \code{"list"}}

\item{geometry}{GeoJSON geometry object (Point, LineString or Polygon)}

\item{...}{Optional request attributes as described \href{https://openrouteservice.org/dev/#/api-docs/pois}{here}}

\item{api_key}{Character scalar containing openrouteservice API key}

\item{output}{Output format. By default the response is being parsed to a list-based R object}
}
\value{
A list of points of interest in the area specified in \code{geometry} structured according to \code{output}:

\itemize{
\item for \code{"text"}, a character vector of length 1 re-encoded to UTF-8.
}

\itemize{
\item for \code{"parsed"}, a parsed R object.
}
}
\description{
Search for points of interest around points or in geometries.
}
\details{
There are three different request types: \code{pois}, \code{stats} and \code{list}.

\code{pois} returns a GeoJSON FeatureCollection in the specified bounding box or
geometry. \code{stats} does the same but groups by categories, ultimately
returning a JSON object with the absolute numbers of POIs of a certain group.

\code{list} returns a list of category groups and their ids.
}
\examples{
ors_pois('list')
}
\author{
Andrzej Oleś <\email{andrzej@openrouteservice.org}>
}
