% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{ors_profile}
\alias{ors_profile}
\title{Openrouteservice Profiles}
\usage{
ors_profile(
  mode = c("car", "hgv", "bike", "roadbike", "mtb", "e-bike", "walking", "hiking",
    "wheelchair")
)
}
\arguments{
\item{mode}{Profile label.}
}
\value{
Profile name, or named vector of available profiles.
}
\description{
List of available modes of transport.
}
\details{
Convenience function for specifying the profile in \code{\link[=ors_directions]{ors_directions()}},
\code{\link[=ors_isochrones]{ors_isochrones()}} and \code{\link[=ors_matrix]{ors_matrix()}}.
}
\examples{
# list availbale profiles
ors_profile()

# retrieve full profile name based on label
ors_profile("car")
}
\seealso{
\code{\link[=ors_directions]{ors_directions()}}, \code{\link[=ors_isochrones]{ors_isochrones()}}, \code{\link[=ors_matrix]{ors_matrix()}}
}
\author{
Andrzej Oleś \href{mailto:andrzej.oles@gmail.com}{andrzej.oles@gmail.com}
}
