% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call.R
\name{print.ors_api}
\alias{print.ors_api}
\title{Print a Compact Summary of the API Response}
\usage{
\method{print}{ors_api}(x, give.attr = FALSE, list.len = 6L, ...)
}
\arguments{
\item{x}{object of class \code{ors_api}.}

\item{give.attr}{logical; if \code{TRUE} (default), show attributes
    as sub structures.}

\item{list.len}{numeric; maximum number of list elements to display
    within a level.}

\item{...}{further arguments passed to \link[utils:str]{str}.}
}
\value{
\code{print.ors_api} prints its argument and returns it \emph{invisibly}.
}
\description{
\code{print.ors_api} uses \link[utils:str]{str} to compactly display the structure
of the openrouteservice API response object.
}
