% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directions.R
\name{ors_directions}
\alias{ors_directions}
\title{Openrouteservice Directions}
\usage{
ors_directions(coordinates, profile = ors_profile(),
  format = c("geojson", "json", "gpx"), ..., api_key = ors_api_key(),
  parse_output = NULL)
}
\arguments{
\item{coordinates}{List of \code{longitude, latitude} coordinate pairs visited in order, alternatively a two column \code{matrix} or \code{data.frame}.}

\item{profile}{Route profile, defaults to \code{"driving-car"}.}

\item{format}{Response format, defaults to \code{"geojson"}}

\item{...}{Optional parameters as described \href{https://openrouteservice.org/dev/#/api-docs/directions}{here}.}

\item{api_key}{Character scalar containing openrouteservice API key}

\item{parse_output}{Should the response be parsed to an R object}
}
\value{
Route between two or more locations for a selected profile and its
settings as GeoJSON response.
}
\description{
Get directions for different modes of transport.
}
\examples{
coordinates = list(c(8.34234, 48.23424), c(8.34423, 48.26424))

# simple call
ors_directions(coordinates, preference="fastest")

# customized options
ors_directions(coordinates, profile="cycling-mountain", elevation=TRUE)
}
\author{
Andrzej Oleś <\email{andrzej@openrouteservice.org}>
}
