% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones.R
\name{ors_isochrones}
\alias{ors_isochrones}
\title{Openrouteservice Isochrones}
\usage{
ors_isochrones(locations, profile = c("driving-car", "driving-hgv",
  "cycling-regular", "cycling-road", "cycling-safe", "cycling-mountain",
  "cycling-tour", "foot-walking", "foot-hiking"), range = 60, ...,
  parse_output = NULL)
}
\arguments{
\item{locations}{List of \code{longitude, latitude} coordinate pairs}

\item{profile}{Route profile, defaults to \code{driving-car}.}

\item{range}{Maximum range value of the analysis in seconds for time and
meters for distance. Alternatively a comma separated list of specific
single range values.}

\item{...}{Optional parameters as described \href{https://openrouteservice.org/documentation/#/reference/isochrones/isochrones}{here}.}

\item{parse_output}{Should the response be parsed to an R object}
}
\value{
Duration or distance matrix for mutliple source and destination
points.
}
\description{
Obtain areas of reachability from given locations.
}
\details{
The Isochrone Service supports time and distance analyses for one single or
multiple locations. You may also specify the isochrone interval or provide
multiple exact isochrone range values.
}
\examples{
ors_isochrones(c(8.34234, 48.23424), interval=20)
}
\author{
Andrzej Oleś <\email{andrzej@openrouteservice.org}>
}
